refactor: extract WorkspaceState account lifecycle into an extension file
Move bootstrap, login/signup, account selection/switching, removal and
data reset, runtime online/offline, and account view-model resolution into
WorkspaceState+Account.swift. Pure code movement, no behavior change.
Build green; all 208 tests pass.
